Add event sourcing skeleton
This commit introduces some of the boilerplate needed to perform event
sourcing. The SQLStreamStore.FSharp library was chosen to leave the
door for sqlite backend potentially open, or ease the transition to
EventStoreDb, since the apis are similar.

A decision was tenatively made to keep state hydration logic in the
domain and implement `decide` as
`DomainEvent list -> DomainCommand -> Result<DomainEvent list, DomainError>`,
rather than the usual `State -> Command -> Event list` to defer that
implmentation until later.